+++ identpy/Model/__init__.py
@@ -2,8 +2,8 @@
 from identpy.Model.model import Model
 
 # Models
-from identpy.Model.systemSM import SpringMass
-from identpy.Model.systemPend import Pendulum
-from identpy.Model.systemZIM import ZIM
-from identpy.Model.systemDFIG import DFIG
-from identpy.Model.improved_DFIG import improved_DFIG
+from identpy.Model.Spring_Mass import SpringMass
+from identpy.Model.Pendulum import Pendulum
+from identpy.Model.ZIM_linearized import ZIM_linearized
+from identpy.Model.DFIG_Erlich import DFIG_Erlich
+from identpy.Model.DFIG_improved import DFIG_improved
